+SignIn mechanism

+++ script/logout.js
@@ -1,3 +1,16 @@
 function logout(){
-    sessionStorage.clear();
+    if(sessionStorage.Storage.getItem("SignIn") == "1"){
+        document.getElementById("LinkBarTableFirstLink").href="SignIn.html";
+        document.getElementById("LinkBarTableFirstText").innerHTML="Sign In ";
+        document.getElementById("LinkBarTableSecondLink").href="SignUp.html";
+        document.getElementById("LinkBarTableSecondText").innerHTML=" Sign Up";
+        sessionStorage.clear();
+    } else {
+        document.getElementById("LinkBarTableFirstLink").href="SignIn.html";
+        document.getElementById("LinkBarTableFirstText").innerHTML="Sign In ";
+        document.getElementById("LinkBarTableSecondLink").href="SignUp.html";
+        document.getElementById("LinkBarTableSecondText").innerHTML=" Sign Up";
+    
+        document.getElementById("LogoutMainBlockDiv").innerHTML="You must sign in first!"
+    }
 }
